Extract scene editor state hooks
